requirements.txt
beautifulsoup4==4.11.1
google_api_python_client==2.127.0
google_auth_oauthlib==1.2.0
jsonpath_ng==1.6.1
ollama==0.1.9
pandas==1.5.3
protobuf==5.26.1
spacy_llm==0.7.1
tqdm==4.65.0
